convert monolith file into modules, update readme
