+++ README.md
@@ -1,4 +1,4 @@
-## Description
+### Description
 
 This is a review/learning project for Binary Search Trees in JavaScript.
 
@@ -9,3 +9,17 @@ Implementation includes:
 * ...more to come
 
 The goal of this project is practice, recollection, and potentially come up with ideas for practical use.
+
+### How To Use:
+
+* Node 16+ is required to run this
+* Open a terminal and navigate to the project root.
+* `npm run examples`
+* Results for examples will apear on your terminal.
+
+### Use cases/ideas for this algorithm:
+
+* left sum, right sum
+* find & count shortest path to node
+* Animate the formation and traversal of nodes on a browser UI.
+* GUI of a virtual "mouse" solving a maze using BST traversals, with toggles & controls
