+++ classes/node_class.js
@@ -1,0 +1,9 @@
+export class Node {
+    constructor(data, left = null, right = null) {
+        this.data = data
+        this.left = left
+        this.right = right
+    }
+}
+
+export default Node;
